requirement.txt
celery==3.1.23
Django==1.7.4
psycopg2==2.5.4
django-bootstrap3==7.0.1
django-celery==3.1.17
requests==2.8.1
xmltodict==0.10.2
django-pgjson==0.3.1
pytz==2016.4
pycrypto==2.6.1
